Adding Canvas to Window
Adding Canvas to Window

Window will now directly draw the pixels from Canvas, allowing the Renderer to write to the Canvas and Window to read from it.

Removed Buffered Paint as it is no longer needed (and technically still supported with Compatible DC and Bitmap).

Remove custom Bitmap test drawing logic.

Note the Window will stretch the Canvas to the Window size.