Canvas Color issue.
Fixed Canvas to Window Color issue.

The issue was that Bitmaps use BGRA  color format, but I prefer RGBA.

I could have used RGBQuad, which has the color values in the correct order, but Initializer list still wouldn't have worked.

COLORREF was supposed to work? But didn't? I don't know why. It's a DWORD  in the form of ABGR, so maybe that's why?

Instead I use BitmapV4Header which allows for Color Masking, or indicating where each RGB Color is in the Bitmap data. COLORREF now works.

Unless we have a need for additional Color functionality, I'll likely stick with COLORREF for now.